+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Changed
+- ***Breaking Change***: Moved `Trampoline` and `RecursiveResult` to better package
 
 ## [2.1.0] - 2018-01-14
 ### Changed